[NOT UPSTREAMABLE] Make scripts compatible with Python 3
Here are all the changes which are needed to make the script compatible with
Python 3 while keeping compatibility with Python 2.6 and 2.7:

* use sys.maxsize instead of sys.maxint,
* use print as a function,
* use "as" keyword in except statements.

--

There have been an issue on the first submission of this patch, concerning
support/genhomedircon.  This script will be removed in short future.
Please read http://oss.tresys.com/pipermail/refpolicy/2014-March/007027.html
for more details.