[NOT UPSTREAMABLE] Add build-time distribution detection to Makefile
When using the same policy on several hosts with different Linux distributions
the DISTRO variable can't be defined in the build.conf of the build directory
because this file is tracked with git.

Instead of maintaining a local patch per host to define DISTRO in build.conf
or of building using "make DISTRO=...", this commit introduces a DISTRO_DETECT
boolean in build.conf which automatically fills DISTRO with the name of the
distro of the building host.

DISTRO_DETECT definition is not copied in the installed build.conf file
(/usr/share/selinux/refpolicy/include/build.conf) because this file would have
the DISTRO variable set.

THIS PATCH HAS BEEN REJECTED BY UPSTREAM
http://oss.tresys.com/pipermail/refpolicy/2014-March/006969.html