[NOT UPSTREAMABLE] Build: allow using tmp/ as a symlink to a temp dir
As tmp/ folder contains temporary files, it makes sense to put it on a
tmpfs partition.  The official/supported way of doing this is my
specifying "tmpdir" variable to "make", but this is quite heavy to
always specify it.

Detect when the to-be-used tmpdir is a symlink and in such a case, use
the linked directory instead of tmpdir.  This way, by creating "tmp" as
a symlink to "/tmp/refpolicy-tmp" the temporary directory will lie in
/tmp filesystem.

As this modification complicates the build system, it is not intended to
be ever sent upstream.