Support systems with a single /usr/bin directory
On systems such as Arch Linux, all programs which are usually located in
/bin, /sbin, /usr/bin and /usr/sbin are present in /usr/bin and the
other locations are symbolic links to this directory.  With such a
configuration, the file contexts which define types for files in
/bin, /sbin and /usr/sbin need to be duplicated to provide definitions
for /usr/bin/...

As the "/bin vs. /usr/bin" part of the needed definitions has already
been done with the "usr merge" patches, the next step consists in
duplicating file contexts for /usr/sbin.  There are several ways of
doing so:

1/ Introduce file contexts for /usr/bin files which have file contexts
in /usr/sbin.  This is quite difficult to maintain in the long run.

2/ Modify the file contexts at build time so that patterns which begin
with /usr/sbin are modified to /usr/bin/...  This does not work in
practise because the tools do not support duplicate file contexts
definition for the same path pattern, and because refpolicy defines
/usr/bin/imapd as courier_pop_exec_t in courier.fc and /usr/sbin/imapd
as imapd_exec_t in uwimap.fc (is this a bug? It is the only conflict I
got).

3/ Replace /usr/sbin by /usr/s?bin everywhere.  This worked fine but was
not really nice-looking/human-friendly.  This also has been recently
broken by commit 2e7553db639b ("Create / to /usr equivalence for bin,
sbin, and lib, from Russell Coker."), which replaced
/usr/(.*/)?bin(/.*)? with /usr/bin(/.*)?, thus creating a pattern with a
higher priority than /usr/s?bin to define the type of files in /usr/bin.
In order to make this work again, this part needs to be reverted.

4/ Modify the file contexts at build time so that patterns which begin
with /usr/sbin are modified to /usr/s?bin/...  I have been using
/usr/s?bin patterns for several years without experiencing issues so
this would work, with the same fix as described in the previous option.

5/ Move everything under /usr/bin and introduce "/usr/sbin /usr/bin" in
config/file_contexts.subs_dist.

As option 4 is the simplest to implement, do this when the new
SINGLE_USR_BIN variable is activated in build.conf.